+++ crossdock/main.go
@@ -21,17 +21,36 @@
 package main
 
 import (
+	"fmt"
 	"net/http"
+	"os/exec"
 	"sync"
+	"time"
 
 	"github.com/crossdock/crossdock-go"
+	"github.com/uber/tchannel-go"
 	"go.uber.org/zap"
 
 	"github.com/uber/jaeger/crossdock/services"
+
+	"golang.org/x/net/context"
 )
 
 const (
 	behaviorEndToEnd = "endtoend"
+
+	collectorService = "Collector"
+	agentService     = "Agent"
+	queryService     = "Query"
+
+	cmdDir       = "/go/cmd/"
+	collectorCmd = cmdDir + "jaeger-collector %s &"
+	agentCmd     = cmdDir + "jaeger-agent %s &"
+	queryCmd     = cmdDir + "jaeger-query %s &"
+
+	collectorHostPort = "localhost:14267"
+	agentURL          = "http://test_driver:5778"
+	queryServiceURL   = "http://127.0.0.1:16686"
 )
 
 var (
@@ -67,11 +86,11 @@ func main() {
 func (h *clientHandler) initialize() {
 	services.InitializeStorage(logger)
 	logger.Info("Cassandra started")
-	services.InitializeCollector(logger)
+	InitializeCollector(logger)
 	logger.Info("Collector started")
-	agentService := services.InitializeAgent("", logger)
+	agentService := InitializeAgent("", logger)
 	logger.Info("Agent started")
-	queryService := services.NewQueryService("", logger)
+	queryService := NewQueryService("", logger)
 	logger.Info("Query started")
 	traceHandler := services.NewTraceHandler(queryService, agentService, logger)
 	h.Lock()
@@ -88,4 +107,68 @@ func (h *clientHandler) isInitialized() bool {
 	h.RLock()
 	defer h.RUnlock()
 	return h.initialized
+}
+
+// InitializeCollector initializes the jaeger collector
+func InitializeCollector(logger *zap.Logger) {
+	cmd := exec.Command("/bin/bash", "-c",
+		fmt.Sprintf(collectorCmd, "-cassandra.keyspace=jaeger -cassandra.servers=cassandra -cassandra.connections-per-host=1"))
+	if err := cmd.Run(); err != nil {
+		logger.Fatal("Failed to initialize collector service", zap.Error(err))
+	}
+	tChannelHealthCheck(logger, collectorService, collectorHostPort)
+}
+
+// NewQueryService initiates the query service
+func NewQueryService(url string, logger *zap.Logger) *services.QueryService {
+	cmd := exec.Command("/bin/bash", "-c", fmt.Sprintf(queryCmd,
+		"-cassandra.keyspace=jaeger -cassandra.servers=cassandra -cassandra.connections-per-host=1"))
+	if err := cmd.Run(); err != nil {
+		logger.Fatal("Failed to initialize query service", zap.Error(err))
+	}
+	if url == "" {
+		url = queryServiceURL
+	}
+	healthCheck(logger, queryService, url)
+	return services.NewQueryService(url, logger)
+}
+
+// InitializeAgent initializes the jaeger agent.
+func InitializeAgent(url string, logger *zap.Logger) *services.AgentService {
+	cmd := exec.Command("/bin/bash", "-c", fmt.Sprintf(agentCmd,
+		"-collector.host-port=localhost:14267 -processor.zipkin-compact.server-host-port=test_driver:5775 "+
+			"-processor.jaeger-compact.server-host-port=test_driver:6831 -processor.jaeger-binary.server-host-port=test_driver:6832"))
+	if err := cmd.Run(); err != nil {
+		logger.Fatal("Failed to initialize agent service", zap.Error(err))
+	}
+	if url == "" {
+		url = agentURL
+	}
+	healthCheck(logger, agentService, agentURL)
+	return services.NewAgentService(url, logger)
+}
+
+func healthCheck(logger *zap.Logger, service, healthURL string) {
+	for i := 0; i < 100; i++ {
+		_, err := http.Get(healthURL)
+		if err == nil {
+			return
+		}
+		logger.Warn("Health check failed", zap.String("service", service), zap.Error(err))
+		time.Sleep(100 * time.Millisecond)
+	}
+	logger.Fatal("All health checks failed", zap.String("service", service))
+}
+
+func tChannelHealthCheck(logger *zap.Logger, service, hostPort string) {
+	channel, _ := tchannel.NewChannel("test_driver", nil)
+	for i := 0; i < 100; i++ {
+		err := channel.Ping(context.Background(), hostPort)
+		if err == nil {
+			return
+		}
+		logger.Warn("Health check failed", zap.String("service", service), zap.Error(err))
+		time.Sleep(100 * time.Millisecond)
+	}
+	logger.Fatal("All health checks failed", zap.String("service", service))
 }
