+++ crossdock/services/common.go
@@ -22,39 +22,7 @@ package services
 
 import (
 	"fmt"
-	"net/http"
-	"time"
-
-	"github.com/uber/tchannel-go"
-	"go.uber.org/zap"
-
-	"golang.org/x/net/context"
 )
-
-func healthCheck(logger *zap.Logger, service, healthURL string) {
-	for i := 0; i < 100; i++ {
-		_, err := http.Get(healthURL)
-		if err == nil {
-			return
-		}
-		logger.Warn("Health check failed", zap.String("service", service), zap.Error(err))
-		time.Sleep(100 * time.Millisecond)
-	}
-	logger.Fatal("All health checks failed", zap.String("service", service))
-}
-
-func tChannelHealthCheck(logger *zap.Logger, service, hostPort string) {
-	channel, _ := tchannel.NewChannel("test_driver", nil)
-	for i := 0; i < 100; i++ {
-		err := channel.Ping(context.Background(), hostPort)
-		if err == nil {
-			return
-		}
-		logger.Warn("Health check failed", zap.String("service", service), zap.Error(err))
-		time.Sleep(100 * time.Millisecond)
-	}
-	logger.Fatal("All health checks failed", zap.String("service", service))
-}
 
 func getTracerServiceName(service string) string {
 	return fmt.Sprintf("crossdock-%s", service)
