+++ crossdock/services/tracehandler.go
@@ -38,15 +38,6 @@ import (
 
 const (
 	servicesParam = "services"
-
-	samplerParamKey = "sampler.param"
-	samplerTypeKey  = "sampler.type"
-
-	epsilon = 0.00000001
-)
-
-var (
-	defaultProbabilities = []float64{1.0, 0.001, 0.5}
 )
 
 type traceRequest struct {
@@ -86,59 +77,6 @@ func (h *TraceHandler) EndToEndTest(t crossdock.T) {
 	}
 }
 
-// AdaptiveSamplingTest creates traces by hitting a client service and validates that the
-// sampling probability has changed.
-//
-// The test creates a stream of traces which gets the adaptive sampler processor to start
-// calculating the probability. The test will wait until the sampling rates are calculated
-// before creating a large amount of traces with the hopes that at least one trace
-// will be sampled with the new sampling probability. The test will make sure the
-// new traces were indeed sampled with a calculated probability by checking span tags.
-func (h *TraceHandler) AdaptiveSamplingTest(t crossdock.T) {
-	operation := generateRandomString()
-	request := createTraceRequest(jaeger.SamplerTypeRemote, operation, 10)
-	service := t.Param(servicesParam)
-	h.logger.Info("Starting AdaptiveSampling test", zap.String("service", service))
-
-	if err := h.runTest(service, request, h.adaptiveSamplingTest, validateAdaptiveSamplingTraces); err != nil {
-		t.Errorf("Fail: %s", err.Error())
-	} else {
-		t.Successf("Pass")
-	}
-}
-
-func (h *TraceHandler) adaptiveSamplingTest(service string, request *traceRequest) ([]*ui.Trace, error) {
-	stop := make(chan struct{})
-	go h.createTraces(service, *request, stop)
-	defer close(stop)
-
-	var rate float64
-	var err error
-	for i := 0; i < 20; i++ {
-		// Keep checking to see if the sampling rate has been calculated
-		h.logger.Info(fmt.Sprintf("Waiting for adaptive sampling probabilities, iteration %d out of 20", i+1))
-		rate, err = h.agent.GetSamplingRate(service, request.Operation)
-		if err == nil && !isDefaultProbability(rate) {
-			break
-		}
-		time.Sleep(500 * time.Millisecond)
-	}
-	if err != nil || isDefaultProbability(rate) {
-		return nil, errors.Wrap(err, "failed to retrieve adaptive sampling rate")
-	}
-
-	// Sleep until the clients are guaranteed to get the new sampling rates (they poll the agent every 5 seconds)
-	time.Sleep(7 * time.Second)
-
-	request.Count = 500
-	request.Tags = map[string]string{"adaptive": "sampling"}
-	traces, err := h.createAndRetrieveTraces(service, request)
-	if err != nil {
-		return nil, err
-	}
-	return traces, nil
-}
-
 func (h *TraceHandler) runTest(service string, request *traceRequest, tFunc testFunc, vFunc validateFunc) error {
 	traces, err := tFunc(service, request)
 	if err != nil {
@@ -148,25 +86,6 @@ func (h *TraceHandler) runTest(service string, request *traceRequest, tFunc test
 		return err
 	}
 	return nil
-}
-
-// createTraces creates traces every 2 seconds for a total of 30 seconds. The loop can be terminated prematurely
-// by closing the stop channel.
-func (h *TraceHandler) createTraces(service string, request traceRequest, stop chan struct{}) {
-	ticker := time.NewTicker(2 * time.Second)
-	endTicker := time.NewTicker(30 * time.Second)
-	defer ticker.Stop()
-	defer endTicker.Stop()
-	for {
-		select {
-		case <-ticker.C:
-			createTrace(service, &request)
-		case <-stop:
-			return
-		case <-endTicker.C:
-			return
-		}
-	}
 }
 
 func (h *TraceHandler) createAndRetrieveTraces(service string, request *traceRequest) ([]*ui.Trace, error) {
@@ -230,44 +149,6 @@ func validateTraces(expected *traceRequest, actual []*ui.Trace) error {
 		}
 	}
 	return nil
-}
-
-func validateAdaptiveSamplingTraces(expected *traceRequest, actual []*ui.Trace) error {
-	if err := validateTraces(expected, actual); err != nil {
-		return err
-	}
-	for _, trace := range actual {
-		tags := convertTagsIntoMap(trace.Spans[0].Tags)
-		samplerParam, ok1 := tags[samplerParamKey]
-		samplerType, ok2 := tags[samplerTypeKey]
-		if !ok1 || !ok2 {
-			return fmt.Errorf("%s and %s tags not found", samplerParamKey, samplerTypeKey)
-		}
-		probability, err := strconv.ParseFloat(samplerParam, 64)
-		if err != nil {
-			return fmt.Errorf("%s tag value is not a float: %s", samplerParamKey, samplerParam)
-		}
-		if samplerType != jaeger.SamplerTypeProbabilistic {
-			return fmt.Errorf("%s tag value should be '%s'", samplerTypeKey, jaeger.SamplerTypeProbabilistic)
-		}
-		if isDefaultProbability(probability) {
-			return fmt.Errorf("adaptive sampling probability not used")
-		}
-	}
-	return nil
-}
-
-func isDefaultProbability(probability float64) bool {
-	for _, p := range defaultProbabilities {
-		if floatEquals(p, probability) {
-			return true
-		}
-	}
-	return false
-}
-
-func floatEquals(a, b float64) bool {
-	return (a-b) < epsilon && (b-a) < epsilon
 }
 
 // The real trace has more tags than the tags we sent in, make sure our tags were created
