+++ crossdock/services/tracehandler_test.go
@@ -133,79 +133,3 @@ func TestValidateTracesWithCount(t *testing.T) {
 		}
 	}
 }
-
-func TestValidateAdaptiveSamplingTraces(t *testing.T) {
-	tests := []struct {
-		expected *traceRequest
-		actual   []*ui.Trace
-		errMsg   string
-	}{
-		{&traceRequest{Count: 1}, []*ui.Trace{{}}, "expected 1 span, got 0"},
-		{&traceRequest{Count: 1},
-			[]*ui.Trace{
-				{Spans: []ui.Span{
-					{Tags: []ui.KeyValue{
-						{Key: "key", Type: ui.BoolType, Value: true},
-					}},
-				},
-				},
-			}, "sampler.param and sampler.type tags not found"},
-		{&traceRequest{Count: 1},
-			[]*ui.Trace{
-				{Spans: []ui.Span{
-					{Tags: []ui.KeyValue{
-						{Key: "sampler.param", Type: ui.StringType, Value: "0.0203"},
-					}},
-				},
-				},
-			}, "sampler.param and sampler.type tags not found"},
-		{&traceRequest{Count: 1},
-			[]*ui.Trace{
-				{Spans: []ui.Span{
-					{Tags: []ui.KeyValue{
-						{Key: "sampler.param", Type: ui.StringType, Value: "not_float"},
-						{Key: "sampler.type", Type: ui.StringType, Value: "probabilistic"},
-					}},
-				},
-				},
-			}, "sampler.param tag value is not a float: not_float"},
-		{&traceRequest{Count: 1},
-			[]*ui.Trace{
-				{Spans: []ui.Span{
-					{Tags: []ui.KeyValue{
-						{Key: "sampler.param", Type: ui.StringType, Value: "0.003"},
-						{Key: "sampler.type", Type: ui.StringType, Value: "const"},
-					}},
-				},
-				},
-			}, "sampler.type tag value should be 'probabilistic'"},
-		{&traceRequest{Count: 1},
-			[]*ui.Trace{
-				{Spans: []ui.Span{
-					{Tags: []ui.KeyValue{
-						{Key: "sampler.param", Type: ui.StringType, Value: "0.001"},
-						{Key: "sampler.type", Type: ui.StringType, Value: "probabilistic"},
-					}},
-				},
-				},
-			}, "adaptive sampling probability not used"},
-		{&traceRequest{Count: 1},
-			[]*ui.Trace{
-				{Spans: []ui.Span{
-					{Tags: []ui.KeyValue{
-						{Key: "sampler.param", Type: ui.StringType, Value: "0.02314"},
-						{Key: "sampler.type", Type: ui.StringType, Value: "probabilistic"},
-					}},
-				},
-				},
-			}, ""},
-	}
-	for _, test := range tests {
-		err := validateAdaptiveSamplingTraces(test.expected, test.actual)
-		if test.errMsg == "" {
-			assert.NoError(t, err)
-		} else {
-			assert.EqualError(t, err, test.errMsg)
-		}
-	}
-}
